requirements.txt
pytest~=5.4.3
selenium~=3.141.0
pytest-rerunfailures
pytest-xdist
allure-pytest
PyYAML~=5.4.1